build(docker): Remnant, broken TTY workaround
First of all, this condition seems reversed from the intention of the commit that adds
it, `ci(docker): Workaround interactive output w/o TTY`.  Secondly, given changes since
then it should be a sign of a bug if a `$ docker compose run ...` command ends up
pulling an image, so I'd rather notice it than hide it, though I wish compose provided
an option to forbid pulling.  Thirdly, this is pollution from the downstream `ci`
branch.