build(vcs): Project version accessible from host
I added `setuptools_scm` to the development packages from the misunderstanding that it
needed to be run with the project's package on `sys.path`, but that's not the case.
It's useful to have it accessible "from the outside" so that we can pass the version
"into" the environment where the package is installed.  This is particularly useful in
container images or other contexts where we want to avoid including `./.git/`.