docs(badges): Attempt fix broken CodeCov badge
Odd behavior here.  In both the GitLab and GitHub web UIs, the image is broken on the
`develop` branch but works on the `main` branch.  The image also works on PyPI.  The
reStructuredText is rendered as an image element with some other URL, some sort of
host-specific CDN, for both GitLab and GitHub.  The un-rendered image URL from the
`.. figure:: ` directive always works in the browser.

Further, the `https://codecov.io/...` target URL from the CodeCov settings page that
provides badges points to a broken, un-styled page, but adding the `app.` subdomain,
`https://app.codecov.io/...`, points to the working CodeCov web UI.  Both versions of
the image URL work in the browser.

So lets try the version *without* the `app.` subdomain for the image, and see if that
fixes the rendering on `develop` for GitLab and GitHub, but leave the `app.` subdomain
for the target.