build(py): Improve tox vs make venv management
The root issue here is if we run recipes in parallel with `$ make -j` and those recipes
invoke `$ tox ...` and Tox determines that a venv needs to be recreated, then it can end
up recreating them at the same time and deleting files out from under one another.  The
simplest/cleanest way to avoid that race condition is to initialize the related tox
venvs with `$ tox run --notest ...` before going parallel to let tox use it's own logic
and interfere with it as little as possible.  The only time we can't rely on that is
when the recipes run in parallel make changes that cause Tox to recreate the venvs.
Currently the only case where that might occur is compiling requirements.

Previously, I  tried to  address this  using `./var/log/tox/py###/build.log` targets so
reflect whether the Tox venvs had been initialized.  The problem with that is there's no
prerequisite we can use  to signal when those targets need to be  rebuild.  If we use as
prerequisites  the targets  that are  prerequisites  for compiling  requirements or  the
compiled requirements themselves, then we introduce circular dependencies when compiling
requirements.  Nevermind the fact that all this is trying to duplicate the logic we
depend on Tox to provide.

Compiling requirements, however, only requires that `$ ./.tox/py###/bin/pip-compile` be
installed and working in the Tox venvs.  IOW, it doesn't matter if the rest of the venv
is up to date or working.  So do away with the fragile and hard to understand
generalized `./var/log/tox/py###/build.log` targets and all cases that use them as
prerequisites except for compiling requirements.  Replace them with direct
`./.tox/py###/bin/pip-compile` targets and only use them as prerequisites for compiling
requirements.  For all other parallel execution just use  `$ tox run --notest ...`
before going parallel.

Also, improve the target for installing the package in editable/develop mode.  Move it
into the Tox venv itself so the recipe will be re-run when the Tox venvs are recreated.