build(py): Redundant parallel prerequisite updates
I noticed duplicate `$ git fetch ...` calls running `$ make devel-upgrade`.  The root
cause is that recipe runs other recipes in parallel but doesn't ensure that all the
prerequisites that are in common are build first.  I audited all `$ make -j ...`
commands for this issue but this is the only one.  Fix by ensuring all images are built
first.

It might be marginally more efficient to have the `devel-upgrade` target depend on the
individual prerequisites of the image built target to let the image get built as the
prerequisite of the per-Python version upgrade target, but that would also add a lot of
complexity.