docs(base): Clarify template source organization
In my experience, once an application involves more than 3-4 components, code organized
by type starts to impose a significant tax in terms of time and focus.  Since this tax
applies to the inner iteration loop of development, it quickly amounts to significant
loss in productivity and enjoyment.  It also makes research and debugging more
difficult.  Finally, it discourages isolation and encapsulation.