build(base): Workaround unfound types, add exts
I agree with what I read, I think it was on some Debian mailing list, that executable
scripts on PATH shouldn't have extensions because the language an executable is written
in is an implementation detail.  By leaving extensions off, the same executable can be
replaced with a different implementation without affecting downstream consumers.  This
is what shebangs are for.

Unfortunately, I keep running into different tools I want to use that don't honor
shebangs, require extensions to work and don't provide feasible workarounds.  Time to
make the pragmatic concession and stop fighting tools that are slightly more dumb than
they should be, especially for scripts never intended to be installed on `PATH`.