style(base): Workaround Vale one vocab limitation
Because [Vale treats a script's shebang as a
comment](https://github.com/errata-ai/vale/issues/631#issue-1711040012), our only
options for workarounds are either disabling the rule altogether or adding exceptions
for the words in shebangs.  The latter is preferable because allows us to still use the
rule for the rest of that rule's word list.  [There's only one `Vocab` per
configuration](https://github.com/errata-ai/vale.sh/pull/37#issue-1711539270), however,
so to limit the exception to only code comments and to keep the word in the rule's word
list for markup, use separate Vale configurations for code and markup.