style(base): Add another prose linter, proselint
Seems to have many more rules than the `proselint` package for Vale and indeed found a
few issues that Vale did not.  It is, as Vale describes, not syntax aware so we can only
run it on markup.  For example, we have to disable a rule that shouldn't be applied to
code blocks.

This reverts commit e4acf063755708b0cbb61a7d3a5746a80a2ef3dd.