style(base): Evaluate write-good JS prose linter
Now that we have Vale linting most source code comments, there's very little extra that
this tool catches that Vale didn't already catch.  The few things this tool does catch
seems worth it to me, or at least should be left to the real project to decided not to
use.

I'm skipping evaluation of the `Joblint` JavaScript prose linter as it is unmaintained.