style(base): Evaluate alex JS prose linter
I don't find the changes made in this commit to satisfy this prose linter to be
meaningful or an improvement.  They're all about terminology that may be inconsiderate
in the human context, for example 'emcee' vs 'host', that are probably not inconsiderate
in the technical context, for example 'system' vs 'host'.  It does, however, demonstrate
that this tool catches some things that our usage of Vale did not, so I'm keeping it
around until I have good reason to remove it.