build(base): Avoid breaking host Python install
Current Debian base images raise the following error when trying to install `$ tox`
globally inside the image:

```
 => CACHED [devel 3/4] COPY [ ./build-host/requirements.txt.in, ./build-host/ ]                                                                                                          0.0s
 => ERROR [devel 4/4] RUN --mount=type=cache,target=/root/.cache,sharing=locked     pip3 install -r "./build-host/requirements.txt.in" |         tee -a "/home/project-structure/.local  0.5s
------
 > [devel 4/4] RUN --mount=type=cache,target=/root/.cache,sharing=locked     pip3 install -r "./build-host/requirements.txt.in" |         tee -a "/home/project-structure/.local/var/log/proje
ct-structure-host-install.log":
or OS, by passing --break-system-packages.
------
Dockerfile:88
--------------------
  87 |     # hadolint ignore=DL3042
  88 | >>> RUN --mount=type=cache,target=/root/.cache,sharing=locked \
  89 | >>>     pip3 install -r "./build-host/requirements.txt.in" | \
  90 | >>>         tee -a "${HOME}/.local/var/log/${PROJECT_NAME}-host-install.log"
  91 |
--------------------
ERROR: failed to solve: process "/bin/bash -eu -o pipefail -c pip3 install -r \"./build-host/requirements.txt.in\" |         tee -a \"${HOME}/.local/var/log/${PROJECT_NAME}-host-install.log\
"" did not complete successfully: exit code: 1
make[1]: *** [Makefile:276: build-docker-build] Error 1
```

I'm torn on the right way to handle this as I have been from the first time I tried to
codify bootstrapping tox into a project.  Like `$ nvm`, and I assume Ruby's equivalent,
`$ tox` generally expects to be installed globally.  Developers familiar with it expect
to invoke it simply as `$ tox`, documentation refers to it, and it is nice to use it
directly in the inner loop of development when an issue has been reproduced.  OTOH,
Tox *can* be managed locally and explicitly, and there *are* occasionally issues that
are specific to Tox versions, so there's also a consistency and reproducibility benefit
to doing so.

I think I'm going to opt for managing Tox locally to the checkout because I am often
using more recent Tox features and the official OS packages are a full major version
behind.

I researched ways to get pip run from within a venv to install to `/usr/local/bin/` but
options were non-obvious and it seems fragile even if I could get it working.  Then it
occurred to me that even if I got it working it would leave the actual underlying venv
in an unexpected form.  IOW, it would be surprising not to be able to run `$
./.tox/bootstrap/bin/tox`.