style(base): Loosen prose lint weasel word alerts
I debated adding exceptions for the words I found valid non-weasel uses for, but the
list of words this rule considers to be weasel words is quite large, includes many words
not on any weasel word list I could find online, and I found many non-weasel uses for
such words in existing documentation.  Better to leave this to the judgment of the
writer and make addressing them optional.