build(base): Recommended way of installing tox
Up to now, I've always worked extra hard to keep everything local to the checkout that I
possibly could.  In this case, that meant installing `$ tox` into a separate virtual
environment for each project or checkout.  Such meta tools, including `$ tox` and `$
nvm`, usually expect to be installed globally per user.  Fighting that pattern incurs a
significant maintenance burden and I now think it's just not worth it.  If 2 projects
end up needing different versions of `$ tox` or `$ nvm`, I'm not going to solve that in
this project.