style(base): Less magical conditional expressions
It was embarrassingly recently that I learned that the `$ [ ... ]` syntax is just
"syntactic sugar" and that `$ [` is actually just an "alias" for `$ test` except that it
just ignores the final `]` in its arguments.  Exit status is a fundamental building
block in shell and it also took me embarrassingly long to learn that `$ if ...` uses
exit status as the conditional.  I can understand the impulse to make conditionals in
shell scripts to read more like most other programming languages.  Weighed, however,
against the cost of hiding the way to think about shell and exit status, and the less
explicit, more magical behavior of `$ [ ... ]`, I now prefer using `$ test`.