build(py): Opt into new pip-compile default
Addresses this warning:

    WARNING: --strip-extras is becoming the default in version 8.0.0. To silence this warning, either use --strip-extras to opt into the new default or use --no-strip-extras to retain the existing behavior.

But it will probably also address the root cause of this CI failure:

    ++ git status --porcelain
    + test -n ' M requirements/py38/devel.txt'
    + set +x
    make: *** [Makefile:855: test-clean] Error 1
    Checkout is not clean

The failure was intermittent in CI.  I reproduced it on every re-run of the GitHub
Actions job but could only reproduce it on the first run in GitLab CI, every GitLab CI
re-run succeeded and didn't reproduce the issue.  Locally I could never reproduce it
running `$ pip-compile` directly on my laptop but I was able to reproduce it every time
running the same command in the Docker development image.  I was only ever able to
reproduce it for the Python 3.8 variant in any environment, never for any of the other
supported Python versions.

Here's the change when reproduced locally in the development image:

    On branch py-ci
    Your branch is ahead of 'origin/py-ci' by 1 commit.
      (use "git push" to publish your local commits)

    Changes not staged for commit:
      (use "git add <file>..." to update what will be committed)
      (use "git restore <file>..." to discard changes in working directory)
            modified:   requirements/py38/devel.txt

    --------------------------------------------------
    Changes not staged for commit:
    diff --git i/requirements/py38/devel.txt w/requirements/py38/devel.txt
    index a9f8c1832..5bff83032 100644
    --- i/requirements/py38/devel.txt
    +++ w/requirements/py38/devel.txt
    @@ -94,7 +94,7 @@ platformdirs==3.11.0
         # via
         #   black
         #   pylint
    -prospector[with_everything]==1.10.3
    +prospector[with-everything,with_everything]==1.10.3
         # via project-structure (pyproject.toml)
     pycodestyle==2.11.1
         # via
    no changes added to commit (use "git add" and/or "git commit -a")

So I think I've confirmed that the uncommitted changes seen in CI are not meaningful and
will be resolved by using the `--strip-extras` option.