build(docker): Bake VCS config in build-host image
I considered a few things for the `user.email` settings.  Firstly I realized it should
be possible to use the same `build-host` image for multiple projects and that real
projects should probably be able to use the `build-host` image from this project
template here.  OTOH, the `user.signingkey` is most likely specific to an individual
maintainer for one or more projects and for the purposes of `$ git log` it would be best
for the email of the author of such commits to be specific to that same individual.  I
concluded that the full name should reflect that a `CI Runner` authored such commits for
readability in `$ git log` and that the user before the `@` should match, hence
`runner@`.  Further, the domain for the email should be specific to an individual
maintainer but not to a specific project, hence using `${PROJECT_NAMESPACE}`.

Regarding the TLD of the domain of the `user.email`, I considered correctness according
to various standards and specifications:

- https://serverfault.com/a/937808/91832
- https://datatracker.ietf.org/doc/html/rfc8375
- https://en.wikipedia.org/wiki/.local

But I think the idea behind those is that private DNS servers would respond for domains
under such TLDs which isn't really what this is about.  I don't see any reason the
domain of the email needs to be resolvable, but neither could I find any other criteria
to motivate a choice.  Testing locally and on a dedicated server, I found `localhost` to
be the only resolvable "non-real" TLD and it's meaning corresponds to the meaning of the
domain of `user.email` for commits authored inside a `build-host` container.  Use the
`localhost` TLD until a better argument comes along.