build(docker): Separate per-job set up script
If there's any set up that only needs to be run once per CI job, it's a waste to put
into the container's `ENTRYPOINT`, particularly if it takes significant time like
changing the ownership of a whole checkout and `GIT_DIR`.  Unfortunately, different CI
providers use containers and their `ENTRYPOINT` commands in different ways so we can't
assume one way of doing such set up.  They also manage the checkouts differently, so we
can't count on an artifact in the checkout to indicate if such set up has already been
done for this job.  Therefor, we have to provide such set up separately from the
`ENTRYPOINT` so that it can be used correctly for however each CI provider sets up the
checkout and runs jobs.