test(py): Reproduce packaging issues in hooks/CI
Using a wheel makes `$ tox` runs much faster but the wheels it builds can become
polluted by the contents of `./build/lib/`. The only reason I can think of to use a
non-editable install is to reproduce packaging issues or to speed up the inner loop of
development. Since using a wheel is a little slower and can become polluted, I can't see
any benefit anymore. Give up on reproducing packaging issues in the inner loop of
development and switch to an editable install by default.

For reproducing packaging issues, we want Tox to build a wheel from an `sdist` as `$
python -m build` does. For releasing built packages we need both the `sdist` and the
`wheel`. Tox's `package = sdist`, however, only builds the `*.tar.gz` files and lets `$
pip install` build and install the `*.whl` file. The `$ pip install` builds the `*.whl`
file in a cache directory which is hard to identify:

    py312: 1001 W install_package> python -I -m pip install --force-reinstall --no-deps /home/rpatterson/src/media/prunerr/.tox/.tmp/package/45/prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031.tar.gz [tox/tox_env/api.py:427]
    Processing ./.tox/.tmp/package/45/prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031.tar.gz
      Installing build dependencies ... done
      Getting requirements to build wheel ... done
      Preparing metadata (pyproject.toml) ... done
    Building wheels for collected packages: prunerr
      Building wheel for prunerr (pyproject.toml) ... done
      Created wheel for prunerr: filename=prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031-py3-none-any.whl size=55064 sha256=ecf89c2f1fa05b6030b4c438f13aef2b256b351decf15305df0f89df1b661fb1
      Stored in directory: /home/rpatterson/.cache/pip/wheels/75/65/be/6be586cfaa9b317c0f898739181dc9540e8c1632c7b24d94f8
    Successfully built prunerr
    Installing collected packages: prunerr
      Attempting uninstall: prunerr
        Found existing installation: prunerr 2.0.0b2.dev7+ged3a9f6e.d20231031
        Uninstalling prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031:
          Successfully uninstalled prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031
    Successfully installed prunerr-2.0.0b2.dev7+ged3a9f6e.d20231031

It also builds the wheel for each Python version which is much slower and unnecessary if
the wheel is universal.  Therefor, to reproduce packaging issues use Tox's `package =
external` to run `$ python -m build`, have Tox use the resulting `*.whl` file for all
Python versions.  I confirmed that Tox deletes previously build packages so also use
that same file and the preceding `*.tar.gz` for releases.