fix(docker): Entrypoint user should not change
I ran into this while using Prunerr in a real deployment in a compose project. The
different `${PROJECT_NAME}` variable value from that project was changing the name of
the user and group created by the `ENTRYPOINT` which was changing where the
configuration file was being looked up.

I think it's best for the container to be as immutable as possible, even at
runtime. Neither can I find a use case for changing the user name *inside* then
container.