fix(transmission): Intermittent missing field
I kept running into ``KeyError`` exceptions because ``torrent._fields["downloadDir"]``
was missing. It happened for different torrents and in different parts of the code. I
wouldn't be surprised to learn that there's some something in the ``export`` sub-command
that is the root cause, but I couldn't find it. It sure *feels* intermittent so use a
pragmatic workaround for now.