fix(free-space): Broad except catching file errors
A configuration error with a Docker volume resulted in the following exception when
removing files while freeing space:

    FileNotFoundError: [Errno 2] No such file or directory:
    '/media/Library/incomplete/*'

It turns out the previous use of `except socket.error:` was just an import alias of
`OSError` which is too broad and was catching the above exception as if it were a
connection error. Use the appropriate subclass instead.