fix(transmission): Files root always torrent name
While debugging the `$ prunerr export` sub-command, I found yet another torrent whose
name is different from the basename of the root of all it's files. In this case, the
torrent name has a trailing `.` dot that the root of all it's files does not.

This has been a problem in the past but I've never been exactly clear on what the edge
cases around this are. Before this most recent instance of the issue, from reading my
code, I clearly concluded that the right, or at least best, root basename is from the
file paths, *not* the torrent name.

In this most recent instance, however, it appears that Transmission places the files
under the torrent name and *not* the files root basename, because when Servarr imported
the files the `droppedPath` included the trailing dot. I'm taking that as a clear
indication of how Transmission behaves and this reverts the code changes to use the
files root basename.