fix(transmission): Files root always root basename
An earlier commit message:

    fix(transmission): Files root always torrent name

details a confirmed case where Transmission placed download files under the torrent name
and *not* the files root basename. Now, for that same download item, I discovered that
when *verifying* files, it uses the files root basename and *not* the torrent name.

The only confirmed case of Transmission using the torrent name is where it *initially*
places downloaded files after completing download. That was while working on the `$
prunerr export` sub-command which is intended as a one-off task that can't yield perfect
results and requires manual intervention anyways. The previous code has been working
well for day-to-day usage. Finally, given that verifying uses the files root basename,
I'm guessing that is the way Transmission *intends* to work and that initial placement
of downloaded files under the torrent name is a bug and the exception to the rule.

If someone runs into more cases of this, it might be worth it to create a small torrent
that reproduces this and test which of Transmission's operations use which root path and
then audit when Prunerr should use which. For now, just go back to the behavior that has
been most reliable for day-to-day usage.

This reverts commit 8bea95cff7e177b9299f41561ea1e8891b62dac7.