fix(export): Misunderstood sourceTitle match
I misunderstood what `sourceTitle` is in the Servarr history records. I thought it was
the download item torrent name, as opposed to the torrent's files root basename which
can be different. But it's actually the **indexer** title for the release which is often
if not usually different from the download item torrent name, for example with spaces
instead of dots.

It's also common to have different download items with similar variations in both
torrent name and files root basename, again dots for spaces and vice versa, which
makes a significant number of false positive matches. It's rare, but there are also
torrents with different names from the root basename. Some are typos, for example a
trailing dot in the torrent name that isn't in the root basename. Some are torrents
containing a subset of a larger collection where the torrent name reflects the subset,
for example `Foo Series S02E01-10` and `Foo Series S02E10-20`, but the root basename is
the same for all, for example `Foo Series/Season 02/*`.

All this means we should only match on root basenames **and** the specific files and
abandon all other matches.