fix(servarr): Seeding items stuck in downloads
I encountered an instance where the `'eventType': 'downloadIgnored'` record was
sandwiched between two apparently duplicate `'eventType': 'grabbed'` records but the
download item as not in the Servarr queue. As such, we can't count on the most recent
record not being a `'eventType': 'grabbed'` record to indicate that Servarr has had a
chance to act on the download item.