fix(free-space): Insufficient space notify
The code used to free memory between ``$ prunerr daemon`` loops by clearing download
items kept in instance state. Now there's a use case for storing session data between
loops so that templates can access the previous session. After that it is much more
complicated to free memory because some state needs to be cleared and some needs to be
kept. I'm not convinced this is worth it because it's not a matter of leaks, that memory
will be freed and then re-consumed when the next loop updates everything. So freeing
memory between loops, only frees that memory for at most 60 seconds. But if there's some
future use case that shows this is worth it again, I'm not opposed to restoring that
freeing memory logic.