fix(extras): Fully imported pruned in free-space
The original exception for Servarr extra files was intended to omit false fails from the
`partially-imported:` operation because it was notifying when the only files that would
remain imported were extra files imported by Servarr. That change also meant, however,
that imported extra files were not counted in the `{{ item.imported_portion }}` which in
turn meant that fully imported items were being deleted in the `prune:` operation.

Fix this by limiting the exception for Servarr extra files to the `partially-imported:`
operation.