fix(un-registered): Endless repeat, must remove
/me slaps forehead

Obviously (now), moving an unregistered item doesn't work because the download item is
still un-registered and the `un-regsitered:` operation has to be in the `all:` stage and
as such will be applied to the item over and over until the `prune:` operation finally
deletes it. This means we can offer no "undo" option to the user without some ugly
kludge.