fix(prune): Avoid normal operation H&R failures
I got a few H&R warnings when everything was running normally and successfully as far as
I could tell. IOW, there wasn't any multi-day Transmission downtime or other condition
that I'm aware of that might have caused it and the warned torrents were all within a
day of meeting their trackers' seeding requirements. All this tells me that the margin
needs to be increased.