test(cover): Intermittent daemon loop cover hole
When running the tests for the different Python versions in parallel, sometimes the
inner daemon loop would take long enough that it exceeded even the `0.1` polling time,
leaving an intermittent coverage hole. I was able to reproduce this on my laptop and I
think on my home server used as a GitLab CI/CD runner.

I'm concerned that this poll value may still not be conservative enough, for example for
someone doing local development on a Chromebook or a tablet. Though such developers are
much more likely to use CI instead of running locally. So we'll see if this intermittent
coverage hole re-emerges and deal with it then.