fix(prune): Lingering verified stalled unimported
I discovered that several items had been corrupted and verified but stalled at ~99% with
no seeders. So even when they are no longer imported they are never pruned. This came to
significant wasted disk space over time which means it's common enough that we should
automate handling these cases.

Use only the item's `downloadDir`, not it's `downloading` vs `seeding` *status*, when
filtering items for the `free-space` stage.

Also simulate the best approximate `doneDate` for items that are `downloading` so that
`seconds_since_done` can still be used to approximate tracker seeding requirements. I
confirmed in a few examples that the `doneDate` for some of these `downloading` items
was populated with an accurate value, though I didn't confirm whether that is from the
`$ prunerr export` sub-command or normal Transmission operation.