docs(todo): Add bonus-by-age/seeders prune order
While writing this TODO, I revisited the feature to delete un-imported items in an
optimal order and clarified the use cases.

For the primary use case, perma-seeding whole libraries, most such users prefer to keep
as broad a collection of content available as their storage allows, so optimizing
deletion order has little effect because at any given time the download items that
are *not* imported are small in both number and storage space.

Another use case, however, is to optimize a download client that is dedicated to
building ratio and none or few of whose items are imported in any library at all. In
that use case, deleting in optimal order is effective at tuning the download client to
get as much upload and ratio as possible for the available storage space.