docs(todo): Also notify when sufficient free space
I'm not sure, however, that Prunerr's current approach to notifications is well suited
to the underlying use case here. It already struggles to know when to and when not to
send the insufficient free space notification. For example, as it is now that
notification is sent every time Prunerr restarts or reloads the configuration. The root
issue here is that this is a matter of responding to changes in the download client's
"global" state but Prunerr currently mostly only deals with the state of individual
download items.

Maybe leave this to system monitoring tools?